Always generate a new metric stream for each view an instrument matches (#3148)
