+++ test/OpenTelemetry.Tests/Metrics/MetricAPITest.cs
@@ -28,7 +28,7 @@ using Xunit.Abstractions;
 
 namespace OpenTelemetry.Metrics.Tests
 {
-    public class MetricApiTest
+    public class MetricApiTest : MetricTestsBase
     {
         private const int MaxTimeToAllowForFlush = 10000;
         private static readonly int NumberOfThreads = Environment.ProcessorCount;
@@ -364,279 +364,6 @@ namespace OpenTelemetry.Metrics.Tests
             var metricPoint2 = metric2MetricPoints[0];
             Assert.Equal(1, metricPoint2.GetHistogramCount());
             Assert.Equal(20D, metricPoint2.GetHistogramSum());
-        }
-
-        [Fact]
-        public void DuplicateInstrumentRegistration_WithViews_DuplicateInstruments_DifferentDescription()
-        {
-            var exportedItems = new List<Metric>();
-
-            using var meter = new Meter($"{Utils.GetCurrentMethodName()}");
-            var meterProviderBuilder = Sdk.CreateMeterProviderBuilder()
-                .AddMeter(meter.Name)
-                .AddView("instrumentName", new MetricStreamConfiguration() { Description = "newDescription1" })
-                .AddView("instrumentName", new MetricStreamConfiguration() { Description = "newDescription2" })
-                .AddInMemoryExporter(exportedItems);
-
-            using var meterProvider = meterProviderBuilder.Build();
-
-            var instrument = meter.CreateCounter<long>("instrumentName", "instrumentUnit", "instrumentDescription");
-
-            instrument.Add(10);
-
-            meterProvider.ForceFlush(MaxTimeToAllowForFlush);
-            Assert.Equal(2, exportedItems.Count);
-
-            var metric1 = exportedItems[0];
-            var metric2 = exportedItems[1];
-            Assert.Equal("newDescription1", metric1.Description);
-            Assert.Equal("newDescription2", metric2.Description);
-
-            List<MetricPoint> metric1MetricPoints = new List<MetricPoint>();
-            foreach (ref readonly var mp in metric1.GetMetricPoints())
-            {
-                metric1MetricPoints.Add(mp);
-            }
-
-            Assert.Single(metric1MetricPoints);
-            var metricPoint1 = metric1MetricPoints[0];
-            Assert.Equal(10, metricPoint1.GetSumLong());
-
-            List<MetricPoint> metric2MetricPoints = new List<MetricPoint>();
-            foreach (ref readonly var mp in metric2.GetMetricPoints())
-            {
-                metric2MetricPoints.Add(mp);
-            }
-
-            Assert.Single(metric2MetricPoints);
-            var metricPoint2 = metric2MetricPoints[0];
-            Assert.Equal(10, metricPoint2.GetSumLong());
-        }
-
-        [Fact]
-        public void DuplicateInstrumentRegistration_WithViews_TwoInstruments_ThreeStreams()
-        {
-            var exportedItems = new List<Metric>();
-
-            using var meter = new Meter($"{Utils.GetCurrentMethodName()}");
-            var meterProviderBuilder = Sdk.CreateMeterProviderBuilder()
-                .AddMeter(meter.Name)
-                .AddView((instrument) =>
-                {
-                    return new MetricStreamConfiguration() { Name = "MetricStreamA", Description = "description" };
-                })
-                .AddView((instrument) =>
-                {
-                    return instrument.Description == "description1"
-                        ? new MetricStreamConfiguration() { Name = "MetricStreamB" }
-                        : new MetricStreamConfiguration() { Name = "MetricStreamC" };
-                })
-                .AddInMemoryExporter(exportedItems);
-
-            using var meterProvider = meterProviderBuilder.Build();
-
-            var instrument1 = meter.CreateCounter<long>("name", "unit", "description1");
-            var instrument2 = meter.CreateCounter<long>("name", "unit", "description2");
-
-            instrument1.Add(10);
-            instrument2.Add(10);
-
-            meterProvider.ForceFlush(MaxTimeToAllowForFlush);
-            Assert.Equal(3, exportedItems.Count);
-
-            var metricA = exportedItems[0];
-            var metricB = exportedItems[1];
-            var metricC = exportedItems[2];
-
-            Assert.Equal("MetricStreamA", metricA.Name);
-            Assert.Equal(20, GetAggregatedValue(metricA));
-
-            Assert.Equal("MetricStreamB", metricB.Name);
-            Assert.Equal(10, GetAggregatedValue(metricB));
-
-            Assert.Equal("MetricStreamC", metricC.Name);
-            Assert.Equal(10, GetAggregatedValue(metricC));
-
-            long GetAggregatedValue(Metric metric)
-            {
-                var metricPoints = new List<MetricPoint>();
-                foreach (ref readonly var mp in metric.GetMetricPoints())
-                {
-                    metricPoints.Add(mp);
-                }
-
-                Assert.Single(metricPoints);
-                return metricPoints[0].GetSumLong();
-            }
-        }
-
-        [Fact]
-        public void DuplicateInstrumentRegistration_WithViews_TwoIdenticalInstruments_TwoViews_DifferentTags()
-        {
-            var exportedItems = new List<Metric>();
-
-            using var meter = new Meter($"{Utils.GetCurrentMethodName()}");
-            var meterProviderBuilder = Sdk.CreateMeterProviderBuilder()
-                .AddMeter(meter.Name)
-                .AddView((instrument) =>
-                {
-                    return new MetricStreamConfiguration { TagKeys = new[] { "key1" } };
-                })
-                .AddView((instrument) =>
-                {
-                    return new MetricStreamConfiguration { TagKeys = new[] { "key2" } };
-                })
-                .AddInMemoryExporter(exportedItems);
-
-            using var meterProvider = meterProviderBuilder.Build();
-
-            var instrument1 = meter.CreateCounter<long>("name");
-            var instrument2 = meter.CreateCounter<long>("name");
-
-            var tags = new KeyValuePair<string, object>[]
-            {
-                new("key1", "value"),
-                new("key2", "value"),
-            };
-
-            instrument1.Add(10, tags);
-            instrument2.Add(10, tags);
-
-            meterProvider.ForceFlush(MaxTimeToAllowForFlush);
-
-            Assert.Equal(2, exportedItems.Count);
-            var metric1 = new List<Metric>() { exportedItems[0] };
-            var metric2 = new List<Metric>() { exportedItems[1] };
-            var tag1 = new List<KeyValuePair<string, object>> { tags[0] };
-            var tag2 = new List<KeyValuePair<string, object>> { tags[1] };
-
-            Assert.Equal("name", exportedItems[0].Name);
-            Assert.Equal("name", exportedItems[1].Name);
-            Assert.Equal(20, GetLongSum(metric1));
-            Assert.Equal(20, GetLongSum(metric2));
-            CheckTagsForNthMetricPoint(metric1, tag1, 1);
-            CheckTagsForNthMetricPoint(metric2, tag2, 1);
-        }
-
-        [Fact]
-        public void DuplicateInstrumentRegistration_WithViews_TwoIdenticalInstruments_TwoViews_SameTags()
-        {
-            var exportedItems = new List<Metric>();
-
-            using var meter = new Meter($"{Utils.GetCurrentMethodName()}");
-            var meterProviderBuilder = Sdk.CreateMeterProviderBuilder()
-                .AddMeter(meter.Name)
-                .AddView((instrument) =>
-                {
-                    return new MetricStreamConfiguration { TagKeys = new[] { "key1" } };
-                })
-                .AddView((instrument) =>
-                {
-                    return new MetricStreamConfiguration { TagKeys = new[] { "key1" } };
-                })
-                .AddInMemoryExporter(exportedItems);
-
-            using var meterProvider = meterProviderBuilder.Build();
-
-            var instrument1 = meter.CreateCounter<long>("name");
-            var instrument2 = meter.CreateCounter<long>("name");
-
-            var tags = new KeyValuePair<string, object>[]
-            {
-                new("key1", "value"),
-                new("key2", "value"),
-            };
-
-            instrument1.Add(10, tags);
-            instrument2.Add(10, tags);
-
-            meterProvider.ForceFlush(MaxTimeToAllowForFlush);
-
-            Assert.Single(exportedItems);
-            var metric1 = new List<Metric>() { exportedItems[0] };
-            var tag1 = new List<KeyValuePair<string, object>> { tags[0] };
-
-            Assert.Equal("name", exportedItems[0].Name);
-            Assert.Equal(20, GetLongSum(metric1));
-            CheckTagsForNthMetricPoint(metric1, tag1, 1);
-        }
-
-        [Fact]
-        public void DuplicateInstrumentRegistration_WithViews_TwoIdenticalInstruments_TwoViews_DifferentHistogramBounds()
-        {
-            var exportedItems = new List<Metric>();
-
-            using var meter = new Meter($"{Utils.GetCurrentMethodName()}");
-            var meterProviderBuilder = Sdk.CreateMeterProviderBuilder()
-                .AddMeter(meter.Name)
-                .AddView((instrument) =>
-                {
-                    return new ExplicitBucketHistogramConfiguration { Boundaries = new[] { 5.0, 10.0 } };
-                })
-                .AddView((instrument) =>
-                {
-                    return new ExplicitBucketHistogramConfiguration { Boundaries = new[] { 10.0, 20.0 } };
-                })
-                .AddInMemoryExporter(exportedItems);
-
-            using var meterProvider = meterProviderBuilder.Build();
-
-            var instrument1 = meter.CreateHistogram<long>("name");
-            var instrument2 = meter.CreateHistogram<long>("name");
-
-            instrument1.Record(15);
-            instrument2.Record(15);
-
-            meterProvider.ForceFlush(MaxTimeToAllowForFlush);
-
-            Assert.Equal(2, exportedItems.Count);
-            var metric1 = exportedItems[0];
-            var metric2 = exportedItems[1];
-
-            Assert.Equal("name", exportedItems[0].Name);
-            Assert.Equal("name", exportedItems[1].Name);
-
-            var metricPoints = new List<MetricPoint>();
-            foreach (ref readonly var mp in metric1.GetMetricPoints())
-            {
-                metricPoints.Add(mp);
-            }
-
-            Assert.Single(metricPoints);
-            var metricPoint = metricPoints[0];
-            Assert.Equal(2, metricPoint.GetHistogramCount());
-            Assert.Equal(30, metricPoint.GetHistogramSum());
-
-            var index = 0;
-            var actualCount = 0;
-            var expectedBucketCounts = new long[] { 0, 0, 2 };
-            foreach (var histogramMeasurement in metricPoint.GetHistogramBuckets())
-            {
-                Assert.Equal(expectedBucketCounts[index], histogramMeasurement.BucketCount);
-                index++;
-                actualCount++;
-            }
-
-            metricPoints = new List<MetricPoint>();
-            foreach (ref readonly var mp in metric2.GetMetricPoints())
-            {
-                metricPoints.Add(mp);
-            }
-
-            Assert.Single(metricPoints);
-            metricPoint = metricPoints[0];
-            Assert.Equal(2, metricPoint.GetHistogramCount());
-            Assert.Equal(30, metricPoint.GetHistogramSum());
-
-            index = 0;
-            actualCount = 0;
-            expectedBucketCounts = new long[] { 0, 2, 0 };
-            foreach (var histogramMeasurement in metricPoint.GetHistogramBuckets())
-            {
-                Assert.Equal(expectedBucketCounts[index], histogramMeasurement.BucketCount);
-                index++;
-                actualCount++;
-            }
         }
 
         [Fact]
@@ -1542,109 +1269,6 @@ namespace OpenTelemetry.Metrics.Tests
 
             meterProvider.ForceFlush(MaxTimeToAllowForFlush);
             Assert.Empty(exportedItems);
-        }
-
-        private static void ValidateMetricPointTags(List<KeyValuePair<string, object>> expectedTags, ReadOnlyTagCollection actualTags)
-        {
-            int tagIndex = 0;
-            foreach (var tag in actualTags)
-            {
-                Assert.Equal(expectedTags[tagIndex].Key, tag.Key);
-                Assert.Equal(expectedTags[tagIndex].Value, tag.Value);
-                tagIndex++;
-            }
-
-            Assert.Equal(expectedTags.Count, tagIndex);
-        }
-
-        private static long GetLongSum(List<Metric> metrics)
-        {
-            long sum = 0;
-            foreach (var metric in metrics)
-            {
-                foreach (ref readonly var metricPoint in metric.GetMetricPoints())
-                {
-                    if (metric.MetricType.IsSum())
-                    {
-                        sum += metricPoint.GetSumLong();
-                    }
-                    else
-                    {
-                        sum += metricPoint.GetGaugeLastValueLong();
-                    }
-                }
-            }
-
-            return sum;
-        }
-
-        private static double GetDoubleSum(List<Metric> metrics)
-        {
-            double sum = 0;
-            foreach (var metric in metrics)
-            {
-                foreach (ref readonly var metricPoint in metric.GetMetricPoints())
-                {
-                    if (metric.MetricType.IsSum())
-                    {
-                        sum += metricPoint.GetSumDouble();
-                    }
-                    else
-                    {
-                        sum += metricPoint.GetGaugeLastValueDouble();
-                    }
-                }
-            }
-
-            return sum;
-        }
-
-        private static int GetNumberOfMetricPoints(List<Metric> metrics)
-        {
-            int count = 0;
-            foreach (var metric in metrics)
-            {
-                foreach (ref readonly var metricPoint in metric.GetMetricPoints())
-                {
-                    count++;
-                }
-            }
-
-            return count;
-        }
-
-        private static MetricPoint? GetFirstMetricPoint(List<Metric> metrics)
-        {
-            foreach (var metric in metrics)
-            {
-                foreach (ref readonly var metricPoint in metric.GetMetricPoints())
-                {
-                    return metricPoint;
-                }
-            }
-
-            return null;
-        }
-
-        // Provide tags input sorted by Key
-        private static void CheckTagsForNthMetricPoint(List<Metric> metrics, List<KeyValuePair<string, object>> tags, int n)
-        {
-            var metric = metrics[0];
-            var metricPointEnumerator = metric.GetMetricPoints().GetEnumerator();
-
-            for (int i = 0; i < n; i++)
-            {
-                Assert.True(metricPointEnumerator.MoveNext());
-            }
-
-            int index = 0;
-            var metricPoint = metricPointEnumerator.Current;
-            foreach (var tag in metricPoint.Tags)
-            {
-                Assert.Equal(tags[index].Key, tag.Key);
-                Assert.Equal(tags[index].Value, tag.Value);
-                index++;
-            }
         }
 
         private static void CounterUpdateThread<T>(object obj)
